fix url input in surveyjs
